feat:#5 탐색 피드 썸네일카드 목업 제작
