+++ src/pages/OnlyTest/OnlyTest.jsx
@@ -1,0 +1,7 @@
+//test용 컴포넌트입니다.
+
+import ThumnailCard from "@/components/ThumnailCard";
+
+export default function OnlyTest() {
+  return <ThumnailCard />;
+}
